requirements.txt
constant==0.0.4
discordwebhook==1.0.3
python-dotenv==1.0.1
tweepy==4.14.0
